Fix sql statement for the grants condition
The sql statement that determines if the `GRANT` should be applied
or not for the case where the `object_type` is "ALL TABLES IN SCHEMA"
is missing a crucial part in the WHERE clause.

The `SELECT` substatement should have a condition that links the bove
SELECT statements together so that a correlation can be made and the
comparison actually returns correct values.

Withouth the correlation, the SQL statement will always return true, if
a single table in the schema has the given `${_privilege}` set.